space_engine::engine::core::assets::bevy_api

Trait CollectionLoadingExt

pub trait CollectionLoadingExt {
    // Required methods
    fn load_core_collection<A>(&mut self) -> &mut Self
       where A: AssetCollection;
    fn load_collection<A>(&mut self) -> &mut Self
       where A: AssetCollection;
}

The CollectionLoadingExt trait extends the Bevy App structure with methods for loading asset collections. It provides two methods:

load_core_collection<A: AssetCollection>(&mut self) -> &mut Self: Configures the loading state to load core assets before the loading screen.

load_collection<A: AssetCollection>(&mut self) -> &mut Self: Configures the loading state to load additional asset collections.

Required Methods

fn load_core_collection<A>(&mut self) -> &mut Self
where A: AssetCollection,

fn load_collection<A>(&mut self) -> &mut Self
where A: AssetCollection,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.