space_engine::engine::space::commands

Struct BodyBuilder

pub struct BodyBuilder {
    pub(crate) radius: f64,
    pub(crate) mass: f64,
    pub(crate) name: String,
    pub(crate) color: Color,
    pub(crate) orbiting: Option<String>,
    pub(crate) eccentricity: f64,
    pub(crate) semi_major_axis: f64,
    pub(crate) argument_of_periapsis: f64,
    pub(crate) reversed_orbit: bool,
    pub(crate) mean_anomaly_at_epoch: f64,
    pub(crate) belts: Vec<(f64, f64, Color)>,
}

Builder command for creating a celestial body (planet or moon, used with aliases)

Fields
radius: f64mass: f64name: Stringcolor: Colororbiting: Option<String>eccentricity: f64semi_major_axis: f64argument_of_periapsis: f64reversed_orbit: boolmean_anomaly_at_epoch: f64belts: Vec<(f64, f64, Color)>
Implementations

impl BodyBuilder

pub fn new(name: impl Into<String>) -> BodyBuilder

Creates a new BodyBuilder instance with a specified name and default values for other properties.

pub fn mass(self, mass: f64) -> BodyBuilder

Sets the mass of the celestial body and returns the builder instance.

pub fn radius(self, radius: f64) -> BodyBuilder

Sets the radius of the celestial body and returns the builder instance.

pub fn color(self, color: Color) -> BodyBuilder

Sets the color of the celestial body and returns the builder instance.

pub fn orbiting(self, orbiting: impl Into<String>) -> BodyBuilder

Sets the parent celestial body that this body will orbit and returns the builder instance.

pub fn eccentricity(self, eccentricity: f64) -> BodyBuilder

Sets the eccentricity of the body’s orbit and returns the builder instance.

pub fn semi_major_axis(self, semi_major_axis: f64) -> BodyBuilder

Sets the semi-major axis of the body’s orbit and returns the builder instance.

pub fn argument_of_periapsis(self, argument_of_periapsis: f64) -> BodyBuilder

Sets the argument of periapsis for the body’s orbit and returns the builder instance.

pub fn mean_anomaly_at_epoch(self, mean_anomaly_at_epoch: f64) -> BodyBuilder

Sets the mean anomaly at epoch for the body’s orbit and returns the builder instance.

pub fn reversed_orbit(self, reversed_orbit: bool) -> BodyBuilder

Sets whether the body orbits in a reversed direction and returns the builder instance.